+++ src/actors.rs
@@ -1,7 +1,7 @@
 
 pub mod player;
 pub mod circle;
-pub mod rect;
-pub mod ellipse;
-pub mod triangle;
-pub mod hexagon;
+// pub mod rect;
+// pub mod ellipse;
+// pub mod triangle;
+// pub mod hexagon;